refactor(github-actions-usage): Move queries to own file
